Removed Quote struct and renamed Quote related features in Journal Controller
The Quote file was a remnant of the first iteration of the app, when it
had no database, and just displayed some text from a struct. It's not
needed anymore, so it's gone.

Then I'm renaming everything I can to make more sense within the
context of the app (it's not about quotes, it's a journal)